requirements.txt
langchain_google_genai
langchain_openai
langchain
load-dotenv
pydantic
fastapi
uvicorn[standard]